requirements.txt
flask
flask-cors
tensorflow-cpu==2.15.0
pillow
numpy<2.0.0
gunicorn